replace column types complexity eci

+++ etl/complexity/eci_pipeline.py
@@ -89,12 +89,12 @@ class ComplexityECIPipeline(EasyPipeline):
         db_connector = Connector.fetch("clickhouse-database", open("../conns.yaml"))
         dtype = {
             "eci":      "Float32",
-            "ent_id":   "UInt16",
+            "ent_id":   "UInt8",
             "latest":   "UInt8",
             "level":    "UInt8",
             "mun_id":   "UInt16",
             "time_id":  "UInt32",
-            "zm_id":    "UInt16"
+            "zm_id":    "UInt32"
         }
         load_step = LoadStep(
             "complexity_eci", db_connector, if_exists="drop", pk=["time_id", "latest", "ent_id", "zm_id", "mun_id"], dtype=dtype
